+++ example/spec/syncGranule/SyncGranuleSuccessSpec.js
@@ -1,36 +1,43 @@
 const fs = require('fs');
 const path = require('path');
 const {
+  api: apiTestUtils,
+  buildAndExecuteWorkflow,
+  addProviders,
+  cleanupProviders,
+  addCollections,
+  cleanupCollections,
+  LambdaStep
+} = require('@cumulus/integration-tests');
+const { Collection, Execution } = require('@cumulus/api/models');
+const {
   aws: {
     s3,
     s3ObjectExists
   },
-  stringUtils: { globalReplace },
+  constructCollectionId,
   testUtils: {
     randomString
   }
 } = require('@cumulus/common');
-const { api: apiTestUtils, buildAndExecuteWorkflow, LambdaStep } = require('@cumulus/integration-tests');
-const { Execution } = require('@cumulus/api/models');
 const {
   loadConfig,
   templateFile,
   uploadTestDataToBucket,
-  timestampedTestDataPrefix,
-  deleteFolder,
-  getFilesMetadata
+  createTimestampedTestId,
+  createTestDataPath,
+  createTestSuffix,
+  deleteFolder
 } = require('../helpers/testUtils');
 const {
   setupTestGranuleForIngest,
-  loadFileWithUpdatedGranuleId
+  loadFileWithUpdatedGranuleIdPathAndCollection
 } = require('../helpers/granuleUtils');
-
 const config = loadConfig();
 const lambdaStep = new LambdaStep();
 const workflowName = 'SyncGranule';
 
 const granuleRegex = '^MOD09GQ\\.A[\\d]{7}\\.[\\w]{6}\\.006\\.[\\d]{13}$';
-const testDataGranuleId = 'MOD09GQ.A2016358.h13v04.006.2016360104606';
 
 const outputPayloadTemplateFilename = './spec/syncGranule/SyncGranule.output.payload.template.json';
 const templatedOutputPayloadFilename = templateFile({
@@ -40,36 +47,47 @@ const templatedOutputPayloadFilename = templateFile({
 
 const s3data = [
   '@cumulus/test-data/granules/MOD09GQ.A2016358.h13v04.006.2016360104606.hdf.met',
-  '@cumulus/test-data/granules/MOD09GQ.A2016358.h13v04.006.2016360104606.hdf',
-  '@cumulus/test-data/granules/MOD09GQ.A2016358.h13v04.006.2016360104606_ndvi.jpg'
+  '@cumulus/test-data/granules/MOD09GQ.A2016358.h13v04.006.2016360104606.hdf'
 ];
 
-describe('When the Sync Granules workflow is configured to overwrite data with duplicate filenames', () => {
-  const testDataFolder = timestampedTestDataPrefix(`${config.stackName}-SyncGranuleSuccess`);
+describe('When the Sync Granules workflow is configured to overwrite data with duplicate filenames\n', () => {
+  const testId = createTimestampedTestId(config.stackName, 'SyncGranuleSuccess');
+  const testSuffix = createTestSuffix(testId);
+  const testDataFolder = createTestDataPath(testId);
+
   const inputPayloadFilename = './spec/syncGranule/SyncGranule.input.payload.json';
+
+  const providersDir = './data/providers/s3/';
+  const collectionsDir = './data/collections/s3_MOD09GQ_006';
+  const collection = { name: `MOD09GQ${testSuffix}`, version: '006' };
+  const provider = { id: `s3_provider${testSuffix}` };
+  const newCollectionId = constructCollectionId(collection.name, collection.version);
+
   let inputPayload;
   let expectedPayload;
-  const collection = { name: 'MOD09GQ', version: '006' };
-  const provider = { id: 's3_provider' };
   let workflowExecution;
 
   process.env.ExecutionsTable = `${config.stackName}-ExecutionsTable`;
   const executionModel = new Execution();
+  process.env.CollectionsTable = `${config.stackName}-CollectionsTable`;
+  const collectionModel = new Collection();
 
   beforeAll(async () => {
-    // upload test data
-    await uploadTestDataToBucket(config.bucket, s3data, testDataFolder);
+    // populate collections, providers and test data
+    await Promise.all([
+      uploadTestDataToBucket(config.bucket, s3data, testDataFolder),
+      addCollections(config.stackName, config.bucket, collectionsDir, testSuffix),
+      addProviders(config.stackName, config.bucket, providersDir, config.bucket, testSuffix)
+    ]);
+    await collectionModel.update(collection, { duplicateHandling: 'replace' });
 
     const inputPayloadJson = fs.readFileSync(inputPayloadFilename, 'utf8');
-
     // update test data filepaths
-    const updatedInputPayloadJson = globalReplace(inputPayloadJson, 'cumulus-test-data/pdrs', testDataFolder);
-    inputPayload = await setupTestGranuleForIngest(config.bucket, updatedInputPayloadJson, testDataGranuleId, granuleRegex);
+    inputPayload = await setupTestGranuleForIngest(config.bucket, inputPayloadJson, granuleRegex, testSuffix, testDataFolder);
+    const newGranuleId = inputPayload.granules[0].granuleId;
 
-    const granuleId = inputPayload.granules[0].granuleId;
-    const updatedOutputPayload = loadFileWithUpdatedGranuleId(templatedOutputPayloadFilename, testDataGranuleId, granuleId);
-    // update test data filepaths
-    expectedPayload = JSON.parse(globalReplace(JSON.stringify(updatedOutputPayload), 'cumulus-test-data/pdrs', testDataFolder));
+    expectedPayload = loadFileWithUpdatedGranuleIdPathAndCollection(templatedOutputPayloadFilename, newGranuleId, testDataFolder, newCollectionId);
+    expectedPayload.granules[0].dataType += testSuffix;
 
     workflowExecution = await buildAndExecuteWorkflow(
       config.stackName, config.bucket, workflowName, collection, provider, inputPayload
@@ -77,14 +95,16 @@ describe('When the Sync Granules workflow is configured to overwrite data with d
   });
 
   afterAll(async () => {
-    // Remove the granule files added for the test
-    await deleteFolder(config.bucket, testDataFolder);
-
-    // delete ingested granule
-    await apiTestUtils.deleteGranule({
-      prefix: config.stackName,
-      granuleId: inputPayload.granules[0].granuleId
-    });
+    // clean up stack state added by test
+    await Promise.all([
+      deleteFolder(config.bucket, testDataFolder),
+      cleanupCollections(config.stackName, config.bucket, collectionsDir, testSuffix),
+      cleanupProviders(config.stackName, config.bucket, providersDir, testSuffix),
+      apiTestUtils.deleteGranule({
+        prefix: config.stackName,
+        granuleId: inputPayload.granules[0].granuleId
+      })
+    ]);
   });
 
   it('completes execution with success status', () => {